Refactor `build_proof`

+++ kate/grid/src/dims.rs
@@ -11,22 +11,22 @@ pub struct Extension {
 	/// This means extending the height of the grid by some factor.
 	/// `2` would mean doubling the grid upwards, increasing the height by a factor of
 	/// 2 and multiplying the number of rows by 2
-	pub height_factor: u16,
+	pub height_factor: usize,
 	/// This means extending the width of the grid by some factor.
 	/// `2` would mean doubling the grid sideways, increasing the width by a factor of
 	/// 2 and multiplying the number of columns by 2
-	pub width_factor: u16,
+	pub width_factor: usize,
 }
 
 impl Extension {
-    pub fn height(factor: u16) -> Self {
+    pub fn height(factor: usize) -> Self {
         Self {
             height_factor: factor,
             width_factor: 1,
         }
     }
 
-    pub fn width(factor: u16) -> Self {
+    pub fn width(factor: usize) -> Self {
         Self {
             height_factor: 1,
             width_factor: factor,
